Se modifica el tipo de variable que puede almacenar la clase Entrada (anterior -> int, nuevo -> long). Se agregan algunas funciones auxiliares y la posibilidad de eliminar y modificar entradas en la HashTable. Se agrega la clase Contactos que simula una agenda telefónica. Esta es la aplicación real que puede utilizar una HashTable en donde se almacenan los contactos por Nombre - Teléfono (String, long).

+++ Hash_Table/src/Entrada.java
@@ -4,9 +4,9 @@
 * */
     public class Entrada{
         String llave;
-        int valor;
+    java.lang.Long valor;
 
-        public Entrada(String llave, int valor) {
+        public Entrada(String llave, java.lang.Long valor) {
             this.llave = llave;
             this.valor = valor;
